fix: account receive gas limit
